requirements.txt
streamlit>=1.32.0
openai>=1.12.0
google-analytics-data>=0.18.0
google-api-python-client>=2.118.0
google-auth>=2.28.0
feedparser>=6.0.11
python-dotenv>=1.0.1
certifi>=2024.2.2
